+++ packages/adhere-ui-searchtable/e2e/proEditableCellRowDragSortSearchTable.jsx
@@ -69,6 +69,7 @@ class RowDragSort extends ProSearchEditableCellRowDragSortStateTable {
         $search: {
           type: 'input',
           visible: true,
+          isShowLabelSymbol: false,
         },
         $editable: {
           editable: true,